requirements.txt
asgiref==3.2.7
astroid==2.4.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.3.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.24.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.25.11
wrapt==1.12.1
